requirements.txt
torch
torchvision
torchaudio
zstandard
accelerate==0.26.1
datasets
wandb
deepspeed==0.13.1
absl-py
torchinfo
scikit-learn
matplotlib
seaborn
sentencepiece
triton
functorch
xformers
gradio
peft
transformers
